% Generated by roxygen2 (4.0.0): do not edit by hand
\name{d3}
\alias{d3}
\title{A generic method for displaying an object using d3.}
\usage{
d3(x, ...)
}
\arguments{
\item{x}{object to display}

\item{...}{other arguments passed on to methods}
}
\description{
A generic method for displaying an object using d3.
}

