\name{Sample profiling data}
\docType{data}
\alias{nesting_prof}
\alias{reshape_prof}
\title{Sample profiling datasets}
\description{

These two datasets illustrate the results of running \code{\link{parse_rprof}} on the sample \code{\link{Rprof}} output stored in the samples directory.  The output was generated by the code in \code{samples/generate.r}.

}
\usage{
  nesting_prof
  reshape_prof
}
\references{}
\keyword{datasets}
