% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/profile.r
\docType{data}
\name{sample-data}
\alias{sample-data}
\alias{nesting_prof}
\alias{reshape_prof}
\title{Sample profiling datasets}
\format{a data frame}
\usage{
nesting_prof

reshape_prof
}
\description{
These two datasets illustrate the results of running
\code{\link{parse_rprof}} on the sample \code{\link{Rprof}} output stored in
the samples directory.  The output was generated by the code in
\code{samples/generate.r}.
}
\keyword{datasets}
